+++ public/icons/thinking.webp
